feature(admin-server,auth-server): Refactor to support querying account subscriptions from admin-server
Because:
- We need to be able to display account subscription information in the admin panel.

This Commit:
- Establishes a pattern of extracting base classes in order to facilitate shared routines across work-spaces, which in this case is the admin-server and auth-server.
- Move routines that provide access to stripe subscription information to fxa-shared/payments
- Move routines and types that provide access to iap subscriptions to fxa-shared/payments/iap
- Move common routines for reporting to errors to sentry to fxa-shared/sentry
