+++ packages/fxa-auth-server/lib/payments/configuration/manager.ts
@@ -1,145 +1,40 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-import { Firestore } from '@google-cloud/firestore';
 import { randomUUID } from 'crypto';
 import { Logger } from 'mozlog';
 import { Container } from 'typedi';
-import { TypedCollectionReference } from 'typesafe-node-firestore';
 
 import errors from '../../error';
 import { AppConfig, AuthFirestore, AuthLogger } from '../../types';
+
+import {
+  PaymentConfigManager as PaymentConfigManagerBase,
+  PaymentConfigManagerError,
+} from 'fxa-shared/payments/configuration/manager';
 import { PlanConfig } from 'fxa-shared/subscriptions/configuration/plan';
 import { ProductConfig } from 'fxa-shared/subscriptions/configuration/product';
 import { mergeConfigs } from 'fxa-shared/subscriptions/configuration/utils';
-import { ConfigType } from '../../../config';
 
-export class PaymentConfigManager {
-  private firestore: Firestore;
-  private productConfigDbRef: TypedCollectionReference<ProductConfig>;
-  private planConfigDbRef: TypedCollectionReference<PlanConfig>;
-  private log: Logger;
-  private config: ConfigType;
-  private prefix: string;
-  private cancelProductListener: (() => void) | undefined;
-  private cancelPlanListener: (() => void) | undefined;
-
-  private products: Record<string, ProductConfig> = {};
-  private plans: Record<string, PlanConfig> = {};
-
-  private hasLoaded: boolean = false;
-
+export class PaymentConfigManager extends PaymentConfigManagerBase {
   constructor() {
-    this.config = Container.get(AppConfig);
-    this.prefix = `${this.config.authFirestore.prefix}payment-config-`;
-    this.firestore = Container.get(AuthFirestore);
-    this.log = Container.get(AuthLogger) as any as Logger;
-    this.productConfigDbRef = this.firestore.collection(
-      `${this.prefix}products`
-    ) as TypedCollectionReference<ProductConfig>;
-    this.planConfigDbRef = this.firestore.collection(
-      `${this.prefix}plans`
-    ) as TypedCollectionReference<PlanConfig>;
-
-    this.load();
-    this.startListeners();
+    const config = Container.get(AppConfig);
+    const firestore = Container.get(AuthFirestore);
+    const log = Container.get(AuthLogger) as any as Logger;
+    super(config, firestore, log);
   }
 
   /**
-   * Load all products and plans from firestore.
-   *
-   * Note that this will overwrite any existing products and plans.
+   * Merges two configs together.
+   * @param planConfig
+   * @param productConfig
+   * @returns merged config
    */
-  public async load() {
-    const [productResults, planResults] = await Promise.all([
-      this.productConfigDbRef.select().get(),
-      this.planConfigDbRef.select().get(),
-    ]);
-    productResults.docs.forEach((doc) => {
-      this.products[doc.id] = ProductConfig.fromFirestoreObject(
-        doc.data(),
-        doc.id
-      );
-    });
-    planResults.docs.forEach((doc) => {
-      this.plans[doc.id] = PlanConfig.fromFirestoreObject(doc.data(), doc.id);
-    });
-    this.hasLoaded = true;
-  }
-
-  private async maybeLoad() {
-    this.hasLoaded || (await this.load());
-  }
-
-  /**
-   * Start the listeners for changes to the product and plan configs.
-   *
-   * Note that this will exit the process if it fails to start the listeners.
-   */
-  public startListeners() {
-    this.cancelProductListener = this.productConfigDbRef.onSnapshot(
-      (snapshot) => {
-        snapshot.docChanges().forEach((change) => {
-          if (['added', 'modified'].includes(change.type)) {
-            this.products[change.doc.id] = ProductConfig.fromFirestoreObject(
-              change.doc.data(),
-              change.doc.id
-            );
-          } else {
-            if (change.doc.id in this.products) {
-              delete this.products[change.doc.id];
-            }
-          }
-        });
-      },
-      (err) => {
-        this.log.error('startListener', { err });
-        process.exit(1);
-      }
-    );
-    this.cancelPlanListener = this.planConfigDbRef.onSnapshot(
-      (snapshot) => {
-        snapshot.docChanges().forEach((change) => {
-          if (['added', 'modified'].includes(change.type)) {
-            this.plans[change.doc.id] = PlanConfig.fromFirestoreObject(
-              change.doc.data(),
-              change.doc.id
-            );
-          } else {
-            if (change.doc.id in this.plans) {
-              delete this.plans[change.doc.id];
-            }
-          }
-        });
-      },
-      (err) => {
-        this.log.error('startListener', { err });
-        process.exit(1);
-      }
-    );
-  }
-
-  /**
-   * Stop the listeners for changes to the product and plan configs.
-   */
-  public stopListeners() {
-    this.cancelProductListener?.();
-    this.cancelPlanListener?.();
-  }
-
-  /**
-   * Looks up a Firestore ProductConfig or PlanConfig document id based
-   * on the provided Stripe Product or Plan id.
-   */
-  public async getDocumentIdByStripeId(
-    stripeId: string
-  ): Promise<string | null> {
-    const products = await this.allProducts();
-    const plans = await this.allPlans();
-    const match =
-      products.find((product) => product.stripeProductId === stripeId) ||
-      plans.find((plan) => plan.stripePriceId === stripeId);
-    return match?.id ?? null;
+  protected override mergeConfigs(
+    planConfig: PlanConfig,
+    productConfig: ProductConfig
+  ): PlanConfig {
+    return mergeConfigs(planConfig, productConfig);
   }
 
   /**
@@ -214,14 +109,19 @@ export class PaymentConfigManager {
     return planId;
   }
 
-  public async allProducts() {
-    await this.maybeLoad();
-    return Object.values(this.products);
-  }
-
-  public async allPlans() {
-    await this.maybeLoad();
-    return Object.values(this.plans);
+  /**
+   * Looks up a Firestore ProductConfig or PlanConfig document id based
+   * on the provided Stripe Product or Plan id.
+   */
+  public async getDocumentIdByStripeId(
+    stripeId: string
+  ): Promise<string | null> {
+    const products = await this.allProducts();
+    const plans = await this.allPlans();
+    const match =
+      products.find((product) => product.stripeProductId === stripeId) ||
+      plans.find((plan) => plan.stripePriceId === stripeId);
+    return match?.id ?? null;
   }
 
   /**
@@ -229,14 +129,14 @@ export class PaymentConfigManager {
    * with the plan's.
    */
   getMergedConfig(planConfig: PlanConfig) {
-    const productConfig = this.products[planConfig.productConfigId];
-    if (!productConfig) {
-      throw errors.internalValidationError(
-        'getMergedConfig',
-        planConfig,
-        new Error('ProductConfig does not exist')
-      );
+    try {
+      return super.getMergedConfig(planConfig);
+    } catch (err) {
+      if (err instanceof PaymentConfigManagerError) {
+        throw errors.internalValidationError(err.op, err.data, err.error);
+      } else {
+        throw err;
+      }
     }
-    return mergeConfigs(planConfig, productConfig);
   }
 }
