+++ packages/fxa-auth-server/lib/payments/iap/google-play/types/index.ts
@@ -1,11 +1,5 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-export { PurchaseQueryError, PurchaseUpdateError } from './errors';
-export {
-  DeveloperNotification,
-  NotificationType,
-  SubscriptionNotification,
-  TestNotification,
-} from './notifications';
-export { Purchase, SkuType } from './purchases';
+
+export * from 'fxa-shared/payments/iap/google-play/types';
