+++ packages/fxa-auth-server/lib/sentry.js
@@ -12,6 +12,8 @@ const { ExtraErrorData } = require('@sentry/integrations');
 const verror = require('verror');
 const { ERRNO } = require('./error');
 const {
+  formatMetadataValidationErrorMessage,
+  reportValidationError,
   tagCriticalEvent,
   buildSentryConfig,
   tagFxaName,
@@ -274,51 +276,6 @@ function ignoreErrors(error) {
 
   // Ingore specific error numbers
   return IGNORED_ERROR_NUMBERS.includes(error.errno);
-}
-
-/**
- * Format a Stripe product/plan metadata validation error message for
- * Sentry to include as much detail as possible about what metadata
- * failed validation and in what way.
- *
- * @param {string} planId
- * @param {string | import('@hapi/joi').ValidationError} error
- */
-function formatMetadataValidationErrorMessage(planId, error) {
-  let msg = `${planId} metadata invalid:`;
-  if (typeof error === 'string') {
-    msg = `${msg} ${error}`;
-  } else {
-    msg = `${msg}${error.details
-      .map(({ message }) => ` ${message};`)
-      .join('')}`;
-  }
-  return msg;
-}
-
-/**
- * Report a validation error to Sentry with validation details.
- *
- * @param {*} message
- * @param {string | import('@hapi/joi').ValidationError} error
- */
-function reportValidationError(message, error) {
-  const details = {};
-  if (typeof error === 'string') {
-    details.error = error;
-  } else {
-    for (const errorItem of error.details) {
-      const key = errorItem.path.join('.');
-      details[key] = {
-        message: errorItem.message,
-        type: errorItem.type,
-      };
-    }
-  }
-  Sentry.withScope((scope) => {
-    scope.setContext('validationError', details);
-    Sentry.captureMessage(message, Sentry.Severity.Error);
-  });
 }
 
 module.exports = {
