+++ packages/fxa-shared/package.json
@@ -91,6 +91,7 @@
     "@types/uuid": "^8.3.1",
     "ajv": "^6.12.2",
     "apollo-server": "^2.25.3",
+    "app-store-server-api": "^0.3.0",
     "aws-sdk": "^2.1120.0",
     "buf": "^0.1.1",
     "celebrate": "^10.0.1",
