+++ packages/fxa-shared/payments/iap/google-play/types/index.ts
@@ -1,5 +1,6 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-export * from 'fxa-shared/payments/iap/apple-app-store/subscription-purchase';
+export * from './errors';
+export * from './notifications';
+export * from './purchases';
